fix: koin-compose
