+++ feature/authorization/presentation/src/commonMain/kotlin/io/timemates/app/authorization/ui/afterstart/mvi/AfterStartStateMachine.kt
@@ -2,7 +2,7 @@ package io.timemates.app.authorization.ui.afterstart.mvi
 
 import io.timemates.app.authorization.ui.afterstart.mvi.AfterStartStateMachine.Effect
 import io.timemates.app.authorization.ui.afterstart.mvi.AfterStartStateMachine.Event
-import io.timemates.app.foundation.mvi.AbstractStateMachine
+import io.timemates.app.foundation.mvi.StateMachine
 import io.timemates.app.foundation.mvi.EmptyState
 import io.timemates.app.foundation.mvi.UiEffect
 import io.timemates.app.foundation.mvi.UiEvent
@@ -10,7 +10,7 @@ import io.timemates.sdk.authorization.email.types.value.VerificationHash
 
 class AfterStartStateMachine(
     reducer: AfterStartReducer,
-) : AbstractStateMachine<EmptyState, Event, Effect>(
+) : StateMachine<EmptyState, Event, Effect>(
     reducer = reducer,
     middlewares = emptyList(),
 ) {
