+++ feature/authorization/presentation/src/commonMain/kotlin/io/timemates/app/authorization/ui/configure_account/mvi/ConfigureAccountStateMachine.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.Immutable
 import io.timemates.app.authorization.ui.configure_account.mvi.ConfigureAccountStateMachine.Effect
 import io.timemates.app.authorization.ui.configure_account.mvi.ConfigureAccountStateMachine.Event
 import io.timemates.app.authorization.ui.configure_account.mvi.ConfigureAccountStateMachine.State
-import io.timemates.app.foundation.mvi.AbstractStateMachine
+import io.timemates.app.foundation.mvi.StateMachine
 import io.timemates.app.foundation.mvi.UiEffect
 import io.timemates.app.foundation.mvi.UiEvent
 import io.timemates.app.foundation.mvi.UiState
@@ -13,7 +13,7 @@ import io.timemates.sdk.authorization.sessions.types.Authorization
 class ConfigureAccountStateMachine(
     reducer: ConfigureAccountReducer,
     middleware: ConfigureAccountMiddleware,
-) : AbstractStateMachine<State, Event, Effect>(
+) : StateMachine<State, Event, Effect>(
     reducer = reducer,
     middlewares = listOf(middleware),
 ) {
