+++ foundation/mvi/koin-compose/src/commonMain/kotlin/io/timemates/app/mvi/compose/stateMachine.kt
@@ -1,7 +1,6 @@
 package io.timemates.app.mvi.compose
 
 import androidx.compose.runtime.Composable
-import io.timemates.app.foundation.mvi.AbstractStateMachine
 import io.timemates.app.foundation.mvi.StateMachine
 import io.timemates.app.foundation.mvi.UiEffect
 import io.timemates.app.foundation.mvi.UiEvent
@@ -19,6 +18,6 @@ import org.koin.core.parameter.ParametersDefinition
  * @return The created instance of the state machine.
  */
 @Composable
-expect inline fun <TState : UiState, TEvent : UiEvent, TEffect : UiEffect> stateMachine(
+expect inline fun <TState : UiState, TEvent : UiEvent, TEffect : UiEffect, reified TSM : StateMachine<TState, TEvent, TEffect>> stateMachine(
     noinline parameters: ParametersDefinition? = null,
-): StateMachine<TState, TEvent, TEffect>
+): TSM
