+++ foundation/mvi/koin-compose/src/jvmMain/kotlin/io/timemates/app/mvi/compose/stateMachine.kt
@@ -1,7 +1,6 @@
 package io.timemates.app.mvi.compose
 
 import androidx.compose.runtime.Composable
-import io.timemates.app.foundation.mvi.AbstractStateMachine
 import io.timemates.app.foundation.mvi.StateMachine
 import io.timemates.app.foundation.mvi.UiEffect
 import io.timemates.app.foundation.mvi.UiEvent
@@ -10,8 +9,8 @@ import org.koin.compose.koinInject
 import org.koin.core.parameter.ParametersDefinition
 
 @Composable
-actual inline fun <TState : UiState, TEvent : UiEvent, TEffect : UiEffect> stateMachine(
+actual inline fun <TState : UiState, TEvent : UiEvent, TEffect : UiEffect, reified TSM : StateMachine<TState, TEvent, TEffect>> stateMachine(
     noinline parameters: ParametersDefinition?
-) = koinInject<StateMachine<TState, TEvent, TEffect>>(
+): TSM = koinInject<TSM>(
     parameters = parameters,
 )
